backend_python/pyproject.toml
[tool.poetry]
name = "khazad-dum-back"
version = "0.1.0"
description = ""
authors = ["phineas0fog <evrard974@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.54.1"
uvicorn = "^0.11.5"
sqlalchemy = "^1.3.16"
pyjwt = "^1.7.1"
python-dotenv = "^0.13.0"
psycopg2-binary = "^2.8.5"
passlib = "^1.7.2"
alembic = "^1.4.2"
bcrypt = "^3.1.7"
python-multipart = "^0.0.5"
pendulum = "^2.1.0"
loguru = "^0.4.1"
aiofiles = "^0.5.0"
Pillow = "^7.1.2"
httpx = "^0.12.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
